backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.15.1
sentence-transformers>=2.2.2
pillow>=10.4.0
transformers>=4.35.2
torch>=2.1.1
torchvision>=0.16.1
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.3
requests>=2.31.0
aiofiles>=23.2.1
psutil>=5.9.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
slowapi>=0.1.9
redis>=5.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiocache>=0.12.2
aioredis>=2.0.1
async-lru>=2.0.4
databases[asyncpg]>=0.8.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7
structlog>=23.2.0
orjson>=3.9.10
brotli>=1.1.0
gunicorn>=21.2.0
pyinstrument>=4.6.1
